git-svn-id: svn://192.168.1.10/public/CoffeeMud@11315 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Areas/StdArea.java
@@ -49,7 +49,7 @@ public class StdArea implements Area
 	protected int		derivedAtmo		=Places.ATMOSPHERE_INHERIT;
 	protected int   	climask			=Area.CLIMASK_NORMAL;
 	protected int		derivedClimate	=Places.CLIMASK_INHERIT;
-	protected long  	tickStatus  	=Tickable.STATUS_NOT;
+	protected int	  	tickStatus  	=Tickable.STATUS_NOT;
 	protected long  	expirationDate  =0;
 	protected long  	lastPlayerTime  =System.currentTimeMillis();
 	protected State   	flag			=State.ACTIVE;
@@ -812,7 +812,7 @@ public class StdArea implements Area
 				a.next().executeMsg(myHost,msg);
 	}
 
-	public long getTickStatus(){ return tickStatus;}
+	public int getTickStatus(){ return tickStatus;}
 
 	public boolean tick(final Tickable ticking, final int tickID)
 	{
